+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Bug fixes:
 - Delegating a task could trigger a previously-silent exception.
 - If we edit text that matches an active search in an editor the editor will re-trigger the search so the search highlights
   and match counts are correct.
+- If the user hits cancel on a conversation that's waiting on a tool approval this would leave the tool approval widget
+  on the screen.  Now this is cleaned up.
 
 ## v0.24 (2025-08-31)